backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.31.1
pandas>=2.2.3
sqlalchemy>=2.0.36
python-multipart>=0.0.12
pydantic>=2.11.0
aiofiles>=23.2.1
python-dateutil>=2.9.0
reportlab>=4.0.0
pytest>=8.0.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
